requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
cachetools==6.2.4
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
et_xmlfile==2.0.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.59.1
fqdn==1.5.1
fsspec==2026.1.0
gensim==4.4.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.9
langchain==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.1
MarkupSafe==2.1.4
matplotlib==3.10.5
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
narwhals==2.15.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.8.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==2.3.3
opencv-python==4.13.0.90
openpyxl==3.1.5
orjson==3.10.18
overrides==7.7.0
packaging==23.2
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.3
patsy==1.0.1
pillow==11.3.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==6.33.4
psutil==5.9.8
pure-eval==0.2.2
pyarrow==23.0.0
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.2.3
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2025.2
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
scikit-learn==1.7.2
scipy==1.16.1
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==69.1.1
six==1.16.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.41
stack-data==0.6.3
statsmodels==0.14.5
streamlit==1.53.0
svgling==0.5.0
svgwrite==1.4.3
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.0
threadpoolctl==3.6.0
tinycss2==1.2.1
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.2.0
virtualenv==20.25.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wrapt==2.0.1
zstandard==0.23.0
